Place all packages under the main mei package
