+++ src/main/java/mei/fileaccess/FileWrite.java
@@ -1,6 +1,6 @@
-package fileaccess;
+package mei.fileaccess;
 
-import tasks.Task;
+import mei.tasks.Task;
 
 import java.io.File;
 import java.io.FileWriter;
