+++ src/main/java/mei/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package tasks;
+package mei.tasks;
 
 import java.time.LocalDateTime;
