+++ src/main/java/mei/tasks/Event.java
@@ -1,4 +1,4 @@
-package tasks;
+package mei.tasks;
 
 import java.time.LocalDateTime;
