+++ src/main/java/mei/tasks/TimedTask.java
@@ -1,4 +1,4 @@
-package tasks;
+package mei.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
