+++ src/main/java/mei/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package tasks;
+package mei.tasks;
 
 public class ToDo extends Task {
