+++ src/test/java/manager/TaskManagerTest.java
@@ -1,8 +1,9 @@
 package manager;
 
+import mei.manager.TaskManager;
 import org.junit.jupiter.api.Test;
-import stub.FileStorageStub;
-import tasks.Task;
+import mei.stub.FileStorageStub;
+import mei.tasks.Task;
 
 import java.util.ArrayList;
 import java.util.List;
